Moved f-components to different folder. Restored FImage in datatable's checkbox cells.
