+++ packages/ketchup/src/components/kup-card/kup-card.scss
@@ -4,7 +4,7 @@
 @import 'collapsible/kup-card-collapsible';
 @import 'scalable/kup-card-scalable';
 @import 'standard/kup-card-standard';
-@import '../../utils/components/f-image/f-image.scss';
+@import '../../f-components/f-image/f-image.scss';
 
 :host {
   display: block;
