+++ packages/ketchup/src/components/kup-card/scalable/kup-card-scalable.tsx
@@ -1,5 +1,5 @@
 import { h } from '@stencil/core';
-import { FImage } from '../../../utils/components/f-image/f-image';
+import { FImage } from '../../../f-components/f-image/f-image';
 import { KupCard } from '../kup-card';
 
 export function create1(component: KupCard) {
