+++ packages/ketchup/src/components/kup-checkbox/kup-checkbox.scss
@@ -1,4 +1,4 @@
-@import '../../utils/components/f-checkbox/f-checkbox.scss';
+@import '../../f-components/f-checkbox/f-checkbox.scss';
 
 :host {
   display: block;
