+++ packages/ketchup/src/components/kup-chip/kup-chip.scss
@@ -1,4 +1,4 @@
-@import '../../utils/components/f-chip/f-chip.scss';
+@import '../../f-components/f-chip/f-chip.scss';
 
 :host {
   display: block;
