+++ packages/ketchup/src/components/kup-image/kup-image.scss
@@ -1,4 +1,4 @@
-@import '../../utils/components/f-image/f-image.scss';
+@import '../../f-components/f-image/f-image.scss';
 
 :host {
   display: block;
