+++ packages/ketchup/src/components/kup-image/kup-image.tsx
@@ -14,7 +14,7 @@ import { logLoad, logMessage, logRender } from '../../utils/debug-manager';
 import { imageCanvas } from './canvas/kup-image-canvas';
 import { setThemeCustomStyle, setCustomStyle } from '../../utils/theme-manager';
 import { KupBadge } from '../kup-badge/kup-badge';
-import { FImage } from '../../utils/components/f-image/f-image';
+import { FImage } from '../../f-components/f-image/f-image';
 
 @Component({
     tag: 'kup-image',
