+++ packages/ketchup/src/f-components/f-checkbox/f-checkbox.scss
