+++ packages/ketchup/src/f-components/f-image/f-image.scss
